backend/requirements.txt
# Core dependencies
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
bcrypt==3.2.0
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.0
mongomock==4.1.2
coverage==7.3.2
pytest-cov==4.0.0